# services/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/connections": {
      "get": {
        "operationId": "ConnectionsController_list",
        "parameters": [
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Connections"]
      }
    },
    "/api/v1/connections/catalog": {
      "get": {
        "operationId": "ConnectionsController_catalog",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Connections"]
      }
    },
    "/api/v1/connections/inbox": {
      "get": {
        "operationId": "ConnectionsController_inbox",
        "parameters": [
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Connections"]
      }
    },
    "/api/v1/connections/stats": {
      "get": {
        "operationId": "ConnectionsController_stats",
        "parameters": [
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Connections"]
      }
    },
    "/api/v1/connections/{connectorId}/start": {
      "post": {
        "operationId": "ConnectionsController_start",
        "parameters": [
          {
            "name": "connectorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Connections"]
      }
    },
    "/api/v1/connections/{connectorId}/token": {
      "post": {
        "operationId": "ConnectionsController_attach",
        "parameters": [
          {
            "name": "connectorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "externalId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "pattern": "^[A-Za-z0-9_.-]+$"
                  },
                  "accessToken": {
                    "type": "string",
                    "minLength": 20,
                    "maxLength": 8000
                  }
                },
                "required": ["externalId", "accessToken"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Connections"]
      }
    },
    "/api/v1/connections/{connectorId}/callback": {
      "get": {
        "operationId": "ConnectionsController_callback",
        "parameters": [
          {
            "name": "connectorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Connections"]
      }
    },
    "/api/v1/connections/{id}": {
      "delete": {
        "operationId": "ConnectionsController_disconnect",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["Connections"]
      }
    },
...
  },
  "info": {
    "title": "AI Social OS API",
    "description": "Mọi route đều cần đăng nhập trừ khi ghi rõ là công khai. Route thuộc phạm vi workspace cần thêm header `X-Workspace-Id`; thiếu nó là 400, và trỏ vào workspace không phải của mình là 404 chứ không phải 403 — 403 sẽ xác nhận workspace đó có tồn tại.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}